Home to Trips
Moved the views and urls from home to Trips to make everything more organized
